requirements.txt
Django==3.1.7
keras_nightly==2.5.0.dev2021032900
numpy==1.19.5
tensorflow==2.5.0
djangorestframework==3.12.4
keras==2.4.3
Pillow==8.3.1
